[cmake] Use separate FetchSource's per adapter
While attempting to enable the Level-Zero and OpenCL adapters at the
same time git will report an error due to the OpenCL tag not existing in
the intel/llvm clone use for the Level-Zero adapter. Furthermore, if I
wanted to test separate bug fixes for the CUDA and HIP adapters from
different sources, this would also not be possible.

This patch uses a separate FetchSource for each adapter to allow each
adapters source to be fetched independently of the others and moves
those clones into the build directory. This does result in additional
git clones when enabling multiple adapters and again in multiple build
directories but avoids conflicting remote URL's or multiple build
configurations trampling on each others checked out tags.